Reset the currentNode to the contextNode when evaluating DPath subexpressions
When we execute a subexpression (i.e. CompiledDPath.run(...)) the
currentNode currently remains whatever it last was set to. This means
that if you have an expression like

  { ../../foo[../bar] }

when we execute the ../bar subexpression, its currentNode will be foo.
This is incorrect since ../bar should be executed in the original
context. To fix this, anytime we run a subexpression, reset the
currentNode to the original contextNode. This way any subexpressions
will always be executed in the correct original context. Callers are
still required to save and reset the necessary information (e.g.
currentNode) if they need it.

DFDL-1660