Clean up MarkPool when RepParser child parser fails
If a child parser of a RepParser fails because it threw an SDE, the
RepParser does not discard the marks it has taken, resulting in a
MarkPool leak. The MarkPool leak error also hides the SDE so the cause
of the leak is unclear.

This fixes RepParsers to catch SchemaDefinitionErrors, discard any marks
it has taken, and rethrows the SDE. This prevents MarkPool leaks and
allows the SDE to be displayed to the user.

This also modifies the InfosetWrongNodeType error so that it includes
the name of the wrong node and the type that was expected.

DFDL-1662