Add the ability to provide a TDML infoset from a file
Add an optional "type" attribute to the dfdl:dfdlInfoset element,
having values of either "infoset" (default) or "file". If the value of
the attribute is "infoset", then the children of the dfdlInfoset element
are used as the infoset. This is the existing behavior. If the value of
the attribute is "file", then the contents of dfdlInfoset are treated as
a path to a file whose data is loaded as XML and used as the infoset.
For example:

  <tdml:dfdlInfoset type="file">path/to/infoset.xml</tdml:dfdlInfoset>

The file is found using standard resource resolution.

The type="file" feature is useful for tests that have very large XML
output, such as with binary blobs or files with many repetitions of data
fields. In these cases, it may be more convenient to store the expected
test result (or input infoset for unparsing) in a separate file rather
than inside the TDML test file. This is similar to the type="file"
attribute for tdml:documentPart.

DFDL-1396