Updated versions of most libraries used by daffodil.
Unable to update sbt itself, the native packager plugin for sbt, 
icu, jline, scallop, as those require more extensive code changes than I
was willing to do for this. 

Created a tool UpdateEclipseClasspaths - a app that can be run to update
all the eclipse-project/.classpath files to reflect the changes in
build-sbt when versions are changed.

Also fixed erroneous xsi:schemaLocation in a test file that was causing
CLI tests to fail.

DFDL-1655