Changed validator to open a stream.
This lets it take schema XSD files from the jar package (of a schema)
rather than insisting the XSD are in ordinary files.

This was causing suppression of other errors.

Make sure validation of infoset after parse is only called if there
is in fact an infoset (i.e., there wasn't an error) This was also
masking other errors.

Added tdml:testSuite defaultValidation feature so you can turn on
validation for an entire test suite.

Modified validator so that it works even for full validation for a
schema project which uses a schema packaged in a jar. (For example VMF,
which uses mil-std-2045 by way of a jar.)

DFDL-1628, DFDL-1676