Discriminators now PE on unknown nodes/values instead of SDE.
This is required by DFDL spec. Section 9.5.2 which says:
"Attempts to reference parts of the DFDL Infoset that do not exist are
processing errors." but this is only for discriminators. for other
expression situations, those are RSDE.

See tests 

test_discrimPEnotSDE1
test_assertSDENotPE1,
test_occursCountSDENotPE1
test_discrimPEvalueLength1
test_discrimPEvalueLengthEnclosingParent1

which exercise this functionality.

DFDL-1675