Fix division to avoid arithmetic exceptions for bigDecimal division.
Fixes literal constants like 90.0 being turned into integers also, which
is DFDL-1688

Improved printed representation of infoset elements to not put unknown
length information out, and to include the value representation and
validity if known.

Moved number utility methods into a daffodil-lib Numbers object.

This removes the AsIntConverters object, which was always misnamed.

Added some unit tests for Number utilities that we use in DPath.

Functionality change is that literal values that can be double with no
loss of information, will be double. But if information/precision would
be lost they stay as decimals. This was the intent before, but it was
buggy.

This change enables NATO's link16 J3.5 to round trip.

DFDL-1688