Improvements to testing. Fewer unnecessary warnings.
Also fewer duplicate loads of test suites.

Did not turn on deprecation of direct use of 'new DFDLTestSuite(...)'
because there are 132 places to fix. But it is setup to do this.

Verified that if you have an @AfterClass method in the companion object
of a test class, that this @AfterClass method IS called when JUnit is
finished with a test class. Even though the Runner objects are members
of a scala object, and the scala object itself isn't reclaimed, the
@AfterClass method IS called so that the Runner.reset call can be done
for each runner, which discards the test suite object to avoid memory
leaking. 

DFDL-1300 reopened because of 132 usages of 'new DFDLTestSuite(....)'
instead of Runner(...) with AfterClass method. 

DFDL-1300