Fix dfdl:valueLength/contentLength for some properties
We had very subtle bug where we were missing a ++ when adding which
expressions to check for valueLength/contentLength references. This
meant the list was missing a handful of references, which lead to it not
inserting the correct capture length parsers. Some tests are added to
that should regress if we make this error again.

DFDL-1706