Optimize out unnecessary alignment parsers
This keeps track of the approximate alignment and determines if the
ending alignment of the previous element is a multiple of where the
current element needs to be aligned. All tests pass, showing that the
necessary alignment parsers are inserted. Also, while running all the
tests, ~8400 elements were known to be aligned with this change, whereas
before only ~6100 were known to be aligned previously, showing that many
alignment parsers are optimized out. This should help to improve
performance and reduce the amount of circular dependencies in unparsing
cause by alignment unparsers.

DFDL-648