Discriminators on a direct dispatch branch should apply to an outer point of uncertainty
Our current behavior is that discriminators on a direct dispatch branch
are ignored. The DFDL-WG confirmed that this behavior is not correct,
and that such discriminators should discriminate an outer point of
uncertainty. This removes the push/pop of the discriminator stack for
direct dispatch choices to get this behavior.

DFDL-1716