Add unparsing optimization to calculate absolute bit positions with fixed length elements
When unparsing, there are some cases where an element could create one
or more suspensions where some of the suspensions do not have a known
length. Such is the case with padding and OVC. When this happens, all
following data output streams end up with a relative position. Since so
many DOS's would not know their absolute starting bit position, this can
easily result in deadlocks, especially when alignment and
outputValueCalc properties are involved.

However, if an element has a fixed length (e.g.
dfdl:lengthKind="explicit" and dfdl:length is not an expression), then
it is possible to calculate the absolute bit position of such DOS's.
This performas that calculation, allowing for more DOS's to have an
absolute position and minimize deadlocks.

DFDL-1726