Minimize overhead with performance testing
Creating and using a BufferedWriter has some performance overhead,
especially when measuring the parsing of lots of small files. In an
attempt to get a more accurate reading, the daffodil performance command
now uses a new NullWriter which has minimal overhead, removing that
aspect from the performance loop. This still allocates an Infoset
Outputter inside the performance loop since that is something that an
API user will always need to do. Note that this is still not 100%
accurate to real world performance since a Writer of some sort may still
need to be allocated, but different Writers have different overheads, so
we don't want to measure those.

DFDL-1753