Be more lax when figuring out the git tracking branch
Newer versions of git show the tracking branch as origin/2.0.0, but some
older versions show it as remotes/origin/2.0.0. Our regex was too strict
and so would fail with older versions of git. This fixes that regex to
be a little more relaxed when finding the version of the tracking branch
so that older versions of git still work.