Improve performance of Scala and JDOM Infoset Inputters
The current implementation of these two infoset inputters were just
stubs. They convert the respective data (scala.xml.Node or
org.jdom2.Document) to a string, convert that string to a StringReader,
and then create an XMLTextInfosetInputter to generate the appropriate
events. This conversion to a string has some amount of overhead and
should be avoided. This patch modifies the Scala and JDOM infoset
inputters to step through the data structures and return the appropriate
events, avoiding the overhead related to toString conversion and parsing
of the XML string.

Some basic tests show that the performance improvements can vary pretty
wildy, in some cases seeing only a 10% increase, but in others seeing a
60% increase. Regardless, there is always an increase in performance.

DFDL-1760, DFDL-1761