Modify getLong to use the same code path as getByteArray
We currently have two different ways to get binary data that take into
account bitOrder and byteOrder: getByteArray for hexBinary data, and
long converters for integer data. This means there are two separate code
paths that deal with byte and bit order. However, getByteArray is
generic enough that it could be used to get longs.

This patch removes the converters and uses the same logic as hex binary
to get data. This adds a fillByteArray method, which accepts a
preallocated byte array and fills it using the old getByteArray logic.
getByteArray now allocates an array of the appropriate size and calls
fillByteArray. The getLong methods are modified to use a preallocated
array of 8 bytes and call fillByteArray. Since fillByteArray returns an
array of bytes in bigEndian byte order and most significant bit first
bit order, converting that to a long is very efficient.

There is now one function that deals with all the messy bit/ByteOrder
logic, making it easier to fix potential bugs and to optimize in the
future. Performance-wise this sees maybe a 5% performance boost.
However, the real gain from this is much greater test coverage of
fillByteArray since now all long parsing uses that logic. This revealed
two small bugs in the fillByteOrder logic.

DFDL-1737