Revised daffodil-io module to require passing in a FormatInfo object.
This is just one part in addressing this major issue of daffodil
schema compiler slowness/speed.

It also fixes DAFFODIL-1843 (bit order and OVC interaction bug)

The FormatInfo object is used to obtain the very commonly used format
properties like byteOrder, bitOrder, etc. This eliminates the need to
call setters that set these properties, eliminates the need to
save/restore them on backtracking, and when evaluating unparser
suspensions. Instead, the daffodil runtime just passes an object
(the PState/UState) which contains this information - getting it from
the ERD of an element most likely, or if runtime defined, from an
Evaluatable object which caches the property value on the infoset
element.

The state/finfo has a ref to the processor which must be set so that
FormatInfo can obtain the right ERD/MGRD from that.

THis change eliminates all the "change" processors
(e.g., EncodingChangeParser, ByteOrderChangeParser, and
BitOrderChangeParser and
unparser variants thereof) the insertion of which uses compile-time
context-upward analysis. This should help with increasing sharing and
avoiding duplicated work and data structures in the schema compiler
(DAFFODIL-1444)

There is some overhead in this change - performance impact will have to
be measured. It should reduce backtracking overheads to compensate
perhaps.

Unparser runtime was improved also to remove unneeded DOS splits.

We were splitting the DOS even for evaluations of just expressions -
things that couldn't write to the DOS at all. So we were creating
lots of known zero-length buffering DOS for no reason.

Improved scaladoc and comments to the suspension-related code
areas.

It seems our test coverage of LE_MSBF is thin at the unit test level.

Two TDML tests moved from scala-debug to scala-new as they work now.

Enhance TDML runner for config, defaultConfig, in files or embedded.

Note: Entity replacer/cooker for fill byte (FillByteCooker) isn't
behaving
correctly.  See comments in DAFFODIL-1646.

DAFFODIL-1444, DAFFODIL-1843, DAFFODIL-1868