Support then- and else- expressions with different types
XPath does not require that the then- and else- branches of an
if-expresison return the same type. In fact, using various XPath
testers, I've found that it is completely reasonable for branches to
return completely different types, without any conversion going on. For
example, something like this is perfectly legal:

  if (foo) then 1 else 'false'

This can return either xs:int(1) or xs:string('false'). However, in
Daffodil, since everything is strongly typed, these expressions do need
to be converted to a common type. But that type is determined by the
target type of the if-expression, not a generalized type of the
branches.

This patch modifies IfExpression so that the inherent type of the
if-expression is the same as its target type. That has the effect that
the then- and else- expressions will add converters to convert to the
target type of the if-expression. So in the above example, if the target
type of the if-expression was a string, then the then- expression would
have a string converter to convert the xs:int(1) to a string, and the
else- branch would have no extra converter. So the then- and else-
branches do not really need to have a common type. They just need to
both able to be converted to the target type of the if-expression.

DFDL-1719