CLI: Move left over data check to after data is output and flushed
When using the CLI, we log a left over data message before flushing the
writer. If you output the infoset to stdout (the default) and since
logs go to stderr, and is very easy for the message to end up in the
middle of the infoset. This doesn't break anything since usually people
only read from stdout, but it hides the fact that there was left over
data.

This moves the has left over data check to after we output and flush the
infoset. So if there is left over data and the infoset goes to stdout,
you should see the message much more clearly at the bottom of the
screen. Also changes left over data to be a Warning and to return a
succes code, regardless of left over data or not since it isn't really
an error.

DFDL-1721