Update grammar to yield a not yet implemented error when dfdl:lengthKind='endOfParent' for complex types.
JSON5 tests were inappropriately using 'endOfParent'
and the code wasn't actually recognizing this as
a feature that's not yet implemented.  Changed
JSON5 tests to use 'implicit' and now they pass.

Updated all sections where LengthKind is used to also
contain a case for EndOfParent so that we error.

There should hopefully no longer be a branch where
we don't hit this check.

DFDL-1664