Explicitly set a null namespace in the W3CDOMInfosetOutputter when there is no namespace
When printing out a w3cdom tree, if we do not specify a namespace for an
element (i.e. createElement instead of createElementNS) then the element
is output using the default namesapce. In cases where the element is
actually no namespace, an xmlns="" is sometimes required to prevent the
use of the default namespace. Explicitly setting the namespace to null
when there is no namepsace will tell w3c to add the xmlns="" when
necessary when printed. Note that this change isn't required for the API
to work--the namespace is correctly seen as no namespace regardless.
This incorrect default namespace only occurs with the CLI and converting
the w3cdom to text.

DFDL-1795