Update the performance CLI command to use less memory when unparsing
When unparsing, the performance command currently creates N copies of
the input infoset. For large infosets, this can require a lot of memory,
causing trouble for the garbage collector and limiting how much memory
is available to perform the actual unparse.

This patch modifies the performance command so that it only creates a
single infoset (e.g. scala node, jdom tree, etc.) that is shared among
unique instances of InfosetInputters used during unparse. This does rely
on the shared data to not be mutated by the InfosetInputters, but that
is currently always the case since they just read and create events.

For many large files, this greatly reduces memory during the CLI
performance command, and speeds things up since there isn't as much time
spent in the garbage collector.

Also remove the scala-slow and jdom-slow InfosetInputters/Outputters
since they are undocumented and not needed anymore. We have gathered the
performance statistics we needed from them.

DFDL-1790