Use a ThreadLocal for w3c dom infoset inputters in CLI
w3c Documents are not thread-safe, even for reading. This is because the
implementation lazily evaluates parts of the tree and creates/destroys
various caches while reading. It is still safe to read the same tree
multiple times, just not at the same time. This patch adds a ThreadLocal
variable when creating w3c documents in the CLI, so that each thread has
a unique copy that it can use. This prevents any threading issue while
still keeping the total number of copies to minimum to save memory.

Also add W3CDOMInfosetInputter/Outputter to the Java and Scala API's,
alont with a comment describing the non-thread-safe nature of the
inputter.

DFDL-1795