Ignore all *.sbt files in the root directory except for build.sbt
This is useful for automated build systems to add custom *.sbt files to
add customizations to the build. For example, adding custom sbt
configurations to publish to a local repository. Note that this is
necessary since our sbt configurations makes a release a SNAPSHOT if
there are any untracked files (e.g. custom sbt file) or files with
changes. If we don't ignore this *.sbt files, then it will also build a
SNAPSHOT. This all helps Daffodil build cleanly on a jenkins automated
build and publishing to artifactory.

DFDL-1666