Improve nil checking for infoset outputters
The InfosetOutputters use _.isNilled to determine if an infoset element
is nil. However, _.isNilled can throw an exception if the nilled flag
has not been set (e.g. while debugging). This could lead to thrown
exceptions during the 'info infoset' debugger command. This patch adds a
new isNilled helper function to the InfosetOutputter trait that performs
the correct logic to check if the nilled flag is set and determine if an
element is nilled, and updates all the InfosetOutputters to use this
function.

DFDL-1829