Property Resolution - part of fixing schema compilation
Refactors property resolving onto elementrefs and grouprefs so
as to eliminate use of backrefs when accessing properties.

Reorganized DSOM for model groups and references to them.

Converted OOLAGHost, SchemaComponent etc. down to LocalElementBase
into traits so that I can mix them the way I need to so that
Global elements aren't terms, etc.

Group references are allowed to carry annotations like dfdl:assert. It
is global group definitions that cannot. (They can't have annotations
period. Annotations go on the model group within the global group
definition)

SequenceBase is a ModelGroup is a Term, and is
the base of Sequence and SequenceGroupRef.

SequenceGroupDef is not a SequenceBase. It is GroupDefLike in that
it has children. It is not a Term, because the SequenceGroupRef that
refers to it is the Term.

Groups work exactly analogously to Elements now. For Elements, the Terms
are
LocalElementDecl and ElementRef. A GlobalElementDecl is *not* a Term.

Added package scaladoc for dsom and grammar packages.

Decided not to rename ModelGroup, nor ElementBase this go-round.

Decided not to try to fix precision of diagnostic messages - most will
have the location of the element ref/group ref not that of the
referenced definition now.

DFDL-1855