Fix so query-style paths are allowed if unambiguous at runtime.
This fix done for nato-stanag-5516 latest schema which has choice branches that share an element name, and expressions that then refer to that element.
(That Link16 is currently broken is DAFFODIL-1869)

So /foo/bar if bar has multiple occurrances, but isn't an array, is
going to be accepted for compiling, though with a warning.
At runtime it is going to test for multiple values, and if there
are more than 1 it will issue a runtime SDE.

Someday it would be good to improve analysis to recognize choices so
that these warnings aren't issued in that case.

Infoset changed so we get a Runtime SDE if expression evaluates to more
than one node.

Tests added which insure one gets this warning, and the runtime SDE in
proper situations. See test_queryStyle1 and test_queryStyle2.

Modified TDML runner to check for warnings whether or not we also get
errors. This fixes DAFFODIL-1583.
Tests test_warningWhenExpectingSuccess and
test_warningWhenExpectingError are in the TDML Runner tests.

Modified expression compiler so that one can issue warnings from it, and
they get saved on the associated schema component.

Modified debugger to capture and issue any warnings that come from path
expression compilation at debug time.

Eliminated circularity bug problem caused by asking expression tree for
isError which checks errors for *everything*.

Refactored confusing ExpresssionCompiler class.

Methods are now explicitly named: compileProperty, compileDelimiter, and
compileExpression, with shared helper routines.

DAFFODIL-1869, DAFFODIL-1773, DAFFODIL-1583