Remove unused files and cruft
Over the years many files have been added to the repo that are longer
needed in the repo. This includes the DFDL specificaiton, old test
files, out of date documentation, scripts that have since been replaced
by the CLI, and schemas that have been moved to external repositories.
This removes these files from this repository.

DAFFODIL-1859