Fix unparser interactions of bitOrder with suspensions/splits.
Call ustate.bitOrder after each term unparser.
setPriorBitOrder in a few places where it was missing.

This corrects an oversight where the unparser was not modified to handle
bit order changes properly. Because few data formats change bit orders
in the middle of data, this error went unrecognized until it was
discovered in NACT Link16 data where the NACT envelope is MSBF, but the
Link16 payload is LSBF.

Test bitOrderOVC1 exercises changing bit orders where a bit order change
occurs, on a byte boundary, but alignment fill is inserted before to
fill a partial byte. The alignment fill for the fragment byte must be
LSBF even though the element having that alignment fill is MSBF.

The tests test_ep1 and test_ep2 are also envelope-payload tests with BE MSBF
envelop and LE LSBF payloads.

DAFFODIL-1884