Change names of zip/tar/rpm binary artifacts
Change the names of the zip, tgz, and rpm binary artifacts to include
"apache" and "incubating". Note that this does not affect the naming of
the jar artifacts at all (e.g. daffodil-lib_2.11-2.1.0.jar stays the
same). The names now look like this:

  apache-daffodil-2.1.0-incubating-bin.zip
  apache-daffodil-2.1.0-incubating-bin.tgz
  apache-daffodil-2.1.0.incubating-1.noarch.rpm

Note that the rpm file has a slightly different naming convention due to
restrictions imposed by rpm.

DAFFODIL-1896