Move built-in-formats.xsd to test and the backwards compatibility added for it
A handful of schemas use built-in-formats.xsd as a method to get a
default set of DFDL properties. This used to be in the
edu/illinois/ncsa/daffodil namespace, but was moved to
org/apache/daffodil with the move to Apache incubation. This move would
have broken schemas, so code was added to provide backwards
compatability. This code essentially replaced instances of
edu/illinois/ncsa with org/apache/daffodil, but that could break things
if actual schema files were in edu/illinois/ncsa, which is legit and
does happen.

However, built-in-formats.xsd is intended to only be used for testing
and not actual schemas. The GeneralPurposeFormat.xsd from IBM is
provided for this purpose and should be used instead. So move
built-in-format.xsd to src/test/resources and remove the broken
backwards compatibility hack. Also update sbt subproject dependencies so
test projects depend on the core "test" so that built-in-formats.xsd is
on the test classpath. Also update CLI test infrastructure to modify
DAFFODIL_CLASSPATH to include the daffodil-lib test resource directory
for the same reason.

DAFFODIL-1897