Handle ChoiceGroups correctly in immediatelyEnclosingModelGroup
This one took a while to track down for what ended up being a one line
change.  Before when immediatelyEnclosingModelGroup was called on a
child element of a ChoiceGroup, it would return the enclosing ModelGroup
of the ChoiceGroup, not the ChoiceGroup itself.  This caused issues
when the CoiceGroup's parent would have initiatedContent="yes", as the
child element of the ChoiceGroup would typically not have an initiator
defined. This issue also effected nested sequence groups and this commit
fixes those as well.

This commit fixes one of the issues preventing the IBM-TLOG schema
project from working correctly.

DAFFODIL-1885

Changes made based on review

Use GeneralFormat instead of daffoidlTest1