Enable annotation of an XSD with both DFDL annotations and an unrelated annotation language.
Problem is that like DFDL, this additional annotation language defines
annotation elements that have xsd:attribute definitions in them.

And Daffodil issues SDE on those attribute definitions because DFDL
doesn't use attributes.

But they're not being used to describe data. They're just being used in
annotation elements that Daffodil (or any DFDL processor) can simply
ignore.

To allow import of a full-function XML Schema for an annotation
language, Daffodil looks at schemas for use of the DFDL_NAMESPACE. If
the root element of the schema uses the DFDL_NAMESPACE (e.g., to define
the dfdl prefix), then the file is treated as a DFDL schema. If the
DFDL_NAMESPACE is not used, then the file is not imported. It is
ignored.

Rationalized validation. Schemas no longer validated by SchemaSet, but
centralized into DFDLSchemaFile class.

This changes validation behavior somewhat. Some errors are superceded by
others now, so negative tests need to change somewhat.

Some tests used to work only because of bugs where validation was not
occurring but should have been, or passed because validation was
occurring when it should not have. So more precise use of test runners,
with fewer tests using modes like validateTDMLFile=false or
validateDFDLSchemas=false.

DAFFODIL-1907,  DAFFODIL-1473