Correctly check if an element is required
When determining which elements could appear after an element in an
infoset for unparsing, we need to determine if an element is optional or
not to know if its siblings could also appear next. We use a mixture of
isOptional (for scalars) and isRequiredArrayElment (for arrays) to
determine if an element is optional. However, this logic was actually
not correct and has some edge cases it misses. Instead, we should just
use the isRequired function, which has the correct logic to determine if
an element could exist or not.

DAFFODIL-1910