Explicitly use the Xerces factories that use our custom catalog resolver
The DFDLCatalogResolver is a custom resolver that extends the Xerces
EntityResolver. This means that the SAXParserFactory's and
SchemaFactory's that use this resolver must also be the Xerces versions
since only they know how to use the Xerces EntityResolver correctly.

The newInstance(...) method provides a parameter to state explicitly
which SchemaFactory/SAXParserFactory instance to create so that we can
ensure the Xerces versions are used. However, this appears to use a
different ClassLoader than normal. This has become apparent in cases
where the Daffodil dependencies (included Xerces) are in a "fat jar", in
which case newInstance() fails to find the Xerces classes with a
ClassNotFoundException, presumably because it uses a ClassLoader that
does not understand to look in the fat jar.

So, instead of using newInstance() and worrying about ClassLoaders, this
changes all SchemaFactory and SAXParserFactory instances to directly
instantiate the Xerces version, which will load the classes just like
every other class instantiation.

The downside with this approach is that one cannot swap out to a
different SchemaFactory/SAXParserFactory, but that was not possible
anyways due to our custom EntityResolver that depends on the Xerces
version, so this isn't really a new limitation.

DFDL-1924