Adding value of 'bcd' to binaryCalendarRep
Adds parsing and unparsing capability for binary representation of
xs:date, xs:time, and xs:dateTime with a binaryCalendarRep of 'bcd',
including known, runtime determined, and delimited lengths of data.

This uses a combination of the existing BCD parser/unparser and
text calendar parser/unparser, with some added checks for binary
calendars.

DAFFODIL-1941