Fix modulus problem with ULong
The ULong code incurrectly calculates the modulus (remainder) of values
and x%y can have values in the range 0<=val<2y instead of 0<=val<y, due
to the way unsigned arithmetic is performed using signed primitives.

This commit fixes the underlying problem and adds test cases to check
the behaviour is correct with both small (<2^63) and large (>=2^63)
unsigned longs.

DAFFODIL-1714