Refactoring to support sequence separators properly.
New Combinators:

OrderedSeparatedSequence
OrderedUnseparatedSequence
+ someday eventually Unordered versions of both the above.

These combinators each create a Parser and an Unparser, so:

OrderedSeparatedSequenceParser
OrderedSeparatedSequenceUnparser
OrderedUnseparatedSequenceParser
OrderedUnseparatedSequenceUnparser

The OrderedUnseparated flavors have no change to their function.

The OrderedSeparatedParser flavor does not *YET* have change to its
function. It needs to be changed to properly orchestrate separators. Its
arguments will not be a vector of terms (only), but will include other
things allowing it to know if the child term is for an array or scalar,
what the separatorSuppressionPolicy is, the separatorPosition, etc. It
will be passed the grammar term for the separator separately so that it
can orchestrate when the separator is parsed.

A number of failures like test_scenario1_7_postfix.

Successful parse of a zero-length empty optional.

HexBinary or String can be empty valued. The infoset element is
constructed successfully.

The Rep parser has to detect that this is success, but zero-length, and
optional. Therefore is missing, and remove the element from the infoset.

Right now that isn't happening.

daffodil-core-unittest runs clean.

daffodil-test 44 failures.

Many things associated with separators are not working properly. This is
a checkpoint before doing some further major refactoring (I suspect),
and also rebase onto master.